bacлутв-analytics / Added endpoint for saving analytics

+++ backend-analytics/Infrastructure/Helpers/IntegerExtensions.cs
@@ -2,7 +2,7 @@
 {
     internal static class IntegerExtensions
     {
-        public static int RoundOff(this int i)
+        public static int RoundOff(this double i)
         {
             return ((int)Math.Round(i / 10.0)) * 10;
         }
